Add files via upload
I designed a game using scratch called Star Wars, it is a shooting game. A simple explanation is to use the left and right arrow keys to control the movement of the rocket, to avoid enemies” hit，and to use the spacebar to shoot the enemies, scoring one point for each enemy shot down, and failing if the enemy hits you.